extend mock deployer
A new field in the mock configuration was added: initialPhase
It sets the phase of the deploy item status, but only if the phase is either 'Init' or empty.
Furthermore, it suppresses the deploy item being set to 'Succeeded' after a successful reconcile.

The reasoning behind this is that for testing the deploy item timeout, the deploy item needs to be in a non-final phase,
e.g. 'Progressing' for some time. However, if the mock deploy item's 'phase' field is used, the mock deployer will
immediately overwrite any 'Failed' state set by the landscaper again, which makes it hard to observe the timeout behaviour.