refactor templating
This is unfortunately a bigger commit which is hard to split up. It contains the following changes:

1. Templating Refactoring
Templating logic contained multiple methods which were basically identical for the different templates (subinstallations, exports, ...).
This has been consolidated into a single method that is reused for all templates.

2. Import Executions
There is a new field for import executions in the blueprint. Similarly to export executions, it allows to describe data transformations
on the imports in the blueprint. Similar to the importDataMappings too, but these are in the installation.

3. Target References in DeployItem Templates
Before, a target reference had to be rendered into the deployitems. This had two major disadvantages:
- the 'usual' case of using an imported target was unnecessarily complicated, as name and namespace had to be extracted from the import
- it was possible to use targets which were not imported
Now, instead of providing name and namespace, only a reference to an imported target is given. This is easier and enforces imports
of targets that should be used.
The old way has been kept for backward compatibility, but is to be removed in the future.
A new local type 'DeployItemSpecification' had to be introduced, because the previously used type 'DeployItemTemplate' can now only be used
pre-rendering (when targets are referenced by import name) and not post-rendering (when targets are referenced by name and namespace) andymore.

4. 'include'
An 'include' function has been added to the go templater. The implementation is the same as the 'include' function in helm templates.