remove dataobject name hashing on root level
Previously, dataobject and target names from exports on root level were hashed to avoid conflicts with manually
created resources. In installations, the prefix '#' could be used to suppress the hashing.
This is opposed to the core concept, that the landscaper doesn't care where an import comes from. Overwriting
manually created resources with exports is not only ok, but explicitly desired.
Therefore, the hashing is now only done for names which are no valid k8s resource names. It is also still
used for imports/exports of nested installations, where it is used to encode the scope.
The '#' notation has been kept to ensure backward compatibility, but it is not required anymore.