Move landscaper installation instructions
As the gettingstarted directory contains exclusively installation
instructions, installation is a more suitable name that allows to find
the instructions more intuitively. Furthermore a dedicated directory
makes sense (instead of e.g. putting it in the guided tour), since this
is quite important information that is referenced from several
locations.