Update Documentation (#896)
* fix faulty example

* Remove manifest-deployer-tutorial

The contents of this tutorial are essentially covered by the guided tour. Furthermore, this tutorial involves
legacy component-cli and legacy landscaper-cli commands and would therefore require updating.

* Move landscaper installation instructions

As the gettingstarted directory contains exclusively installation
instructions, installation is a more suitable name that allows to find
the instructions more intuitively. Furthermore a dedicated directory
makes sense (instead of e.g. putting it in the guided tour), since this
is quite important information that is referenced from several
locations.

* Unify the installation instructions

There should only exist a single document describing the installation of
the landscaper. This makes it easier to keep the documentation clean and
up to date.

* Remove tutorials

For large parts, the content of the tutorials has been superseded by the
guided tour. Again, maintaining redundant documentation causes
unnecessary overhead.

* Remove unfinished and unused documentation

This documentation is unfinished and unused. Besides, most of its
content is a duplicate of the guided tour.

* Remove inline component descriptor from installation scaffold

Inline component descriptors were announced to be deprecated and should
therefore be removed from the documentation.

* Update component descriptor glossary entry

* (run-int-tests)

* Move tutorial resources to corresponding integration tests

A previous commit removed the entire docs/tutorial folder since the
content was deemed to be a duplicate of the guided tour. But the
resources in this directory are used by some integration tests.

* Adjust further references to tutorial resources (run-int-tests)

* (run-int-tests)

* (run-int-tests)

* (run-int-tests)