Upgrade ocm v0.6.0 to v0.8.0
The ocm library version v0.6.0 had a bug where the component descriptor copy command omitted the creation time. Since the copy command is used to in the comparison whether a component descriptor has changed, in the comparison, the creation time is missing and ocm attempts to update the component in the ocm repository even though the library user did not perform any changes.