upgrade ocmlib (#1010)
* Fix make verify bug

The make verify command internally calls the find command with a -max-depth and a -type option. Thereby, it expects the global -max-depth option to be before the -type option.

* Upgrade ocm v0.6.0 to v0.8.0

The ocm library version v0.6.0 had a bug where the component descriptor copy command omitted the creation time. Since the copy command is used to in the comparison whether a component descriptor has changed, in the comparison, the creation time is missing and ocm attempts to update the component in the ocm repository even though the library user did not perform any changes.

* Adjust incompatible changes of upgraded controller-runtime

Upgrading the ocm version also forced an upgrade of the controller-runtime. In the upgraded version of the controller-runtime, the controller manager options do not contain Port and MetricsBindAddress anymore. Since we used the default Port everywhere anyway, Port was removed.

* (run-int-tests)