Initializes and closes ocm context in the controllers reconcile function
So far, the ocm context was not properly closed. Partly, this is because it is not quite clear until what point in the reconcile logic the ocm context is needed. Since we need a new ocm context for each reconcile anyway, we solve this problem by creating the ocm context in the top level reconcile function and deferring its close. To avoid having to adjust a disproportionate amount of function signatures to hand the ocm context down, we bind it to the go context object that is passed around anyway.