Set up a new context object before each test
Due to the change that the ocm context is now passed through the landscaper within the context object (to be able to initialize it directly at the beginning of the reconcile and finalize it at the end without having to pass it through multiple interfaces), several tests have to be adjusted accordingly. While this change might not be functionally relevant for all of the adjusted tests immediately, it prevents unexpected behavior when adjusting those tests in the future.
Furthermore a discard logger is bound to each context to prevent frequent logging messages that it was "unable to fetch logger from context". If logging is needed, with the new setup, the discard logger just has to be exchanged with an actual logger instance.