Fix regex
Although renovate seemed to be able to detect the go versions with the v prefix (v1.2.3), the prefix will not replace a v1.2.3 with v1.2.4 but with 1.2.4 (thus, omit the prefix) if datasource=go (probably since in the module file, the versions are also omitted). But commands such as go install module@v1.2.3 do not work with module@1.2.3.