Renovate for makefile (#1064)
* (run-int-tests)

* Fix regex

Although renovate seemed to be able to detect the go versions with the v prefix (v1.2.3), the prefix will not replace a v1.2.3 with v1.2.4 but with 1.2.4 (thus, omit the prefix) if datasource=go (probably since in the module file, the versions are also omitted). But commands such as go install module@v1.2.3 do not work with module@1.2.3.

* Remove (?:ENV |ARG )? from matchString

This part of the regex only really makes sense for Dockerfile which are not covered by these custom managers.

* Remove go datasource manager

Since renovate uses the re2 regex flavor which does not support positive lookahead, there is no clean way to define another manager that cares about all datasources but go (since you would need the (!?go) lookahead). Thus, use github-releases instead.