ocm config support (#1055)
* Initializes and closes ocm context in the controllers reconcile function

So far, the ocm context was not properly closed. Partly, this is because it is not quite clear until what point in the reconcile logic the ocm context is needed. Since we need a new ocm context for each reconcile anyway, we solve this problem by creating the ocm context in the top level reconcile function and deferring its close. To avoid having to adjust a disproportionate amount of function signatures to hand the ocm context down, we bind it to the go context object that is passed around anyway.

* Set up a new context object before each test

Due to the change that the ocm context is now passed through the landscaper within the context object (to be able to initialize it directly at the beginning of the reconcile and finalize it at the end without having to pass it through multiple interfaces), several tests have to be adjusted accordingly. While this change might not be functionally relevant for all of the adjusted tests immediately, it prevents unexpected behavior when adjusting those tests in the future.
Furthermore a discard logger is bound to each context to prevent frequent logging messages that it was "unable to fetch logger from context". If logging is needed, with the new setup, the discard logger just has to be exchanged with an actual logger instance.

* Add ocm config parameter to registry access factory

I decided to pass in the entire config map as parameter instead of just the data of the config map to make it immediately obvious in the function signature where this data is typically coming from.

* Implement unit tests for ocm config behavior

* Improve error type checking

errors.As is more reliable than the type switch since it also catches the error if it is wrapped.

* Fix missing close

This error existed for quite a while. But it only surfaced now because the tests now check whether the ocm context can be properly closed.

* Add ocm config field to api types

* Implement ocm config map for container deployer

Since the init container has to fetch the ocm components, the ocm config configmap has to be properly synced into the deployer cluster and then mounted into the init containers file system.

* Use the ocm implementation per default

* Implement an integration test for container deployer with ocm config

* Upgrade ocm version

We need a newer version of ocm than the current release (v0.8.0) since we rely on the fix of a bug in the current ocm repository cache implementation.

* Prevent errors due to empty config data

* Fix test

* Reuse make command

Furthermore, this is useful if you import a non release version of the module in the go mod file. For example, the ocm install command will download the binary of a github release, which will not work with a go pseudoversion number.

* Add ocm config map to helm deployer

* Add tests for ocm config related helm deployer functionality

* Add configmaps resource to clusterrole

The container deployer needs to be able to handle configmaps in order to sync the ocm config configmap.

* Run make generate

* Fix open non-existent file

* Fix accidental return

With the construct I used here before, the error declared and assigned (:=) by FileExists had the scope of the if statement. Thus, the following if err != nil check, did not check this error (as it was intended by me) but the error of the Walk(...) function.

---------

Co-authored-by: Achim Weigel <achim.weigel@sap.com>